refactor(jobs): streamline job synchronization logic by integrating jobRepository and agentService
