+++ web-app/src/app/api/sync/jobs/route.ts
@@ -4,12 +4,8 @@ import pTimeout from "p-timeout";
 
 import { getEnvSecrets } from "@/config/env.secrets";
 import { authenticateCronSecret } from "@/lib/auth/utils";
-import {
-  jobsNotFinishedWhereQuery,
-  lockRepository,
-  prisma,
-} from "@/lib/db/repositories";
-import { lockService, syncJob } from "@/lib/services";
+import { jobRepository, lockRepository } from "@/lib/db/repositories";
+import { agentService, lockService } from "@/lib/services";
 import { Lock } from "@/prisma/generated/client";
 
 const LOCK_KEY = "jobs-sync";
@@ -69,15 +65,13 @@ async function jobSync(): Promise<Response> {
 async function syncAllJobs(): Promise<void> {
   const runningDbUpdates: Promise<void>[] = [];
 
-  const jobs = await prisma.job.findMany({
-    where: jobsNotFinishedWhereQuery(),
-  });
+  const jobs = await jobRepository.getJobsNotFinished();
 
   console.info("Syncing", jobs.length, "jobs");
   // Process 5 jobs at a time
   const limit = pLimit(5);
   for (const job of jobs) {
-    runningDbUpdates.push(limit(() => syncJob(job)));
+    runningDbUpdates.push(limit(() => agentService.syncJob(job)));
   }
   try {
     await Promise.allSettled(runningDbUpdates);
